Fix a crash in RootFragment.
On orientation change and app resume, Dalvik expects fragments to have an empty
constructor. Our RootFragment constructor wasn't empty, since we use a bit of a
hack in order to get the nav drawer able to spawn Activities.

So in RootFragment, this commit overloads the constructor with an empty one,
and makes the argument that RootFragment gets be an Option[Int].

In onCreateView(), we check if the RootFragment's layout argument is None or
not. If it's None, we return null, if it's not, we inflate as normal.

Thanks to @danopia for his help in tracking this down.

:hamburger: