Attempt package search without pkgwat-scala.
Right now, Dispatch won't work on Android
( see https://groups.google.com/forum/#!topic/dispatch-scala/pJC5fwec-pA ).

With that in mind, I re-implemented the search functionality of
pkgwat-scala without Dispatch for now, just so that I can start
working on the UI for some of this.

If/when we get Dispatch working, it'll be easy to switch back to it
and a lot of this will become less hacky.