[PIA] Show a bit more package info.
This is mostly to experiment with UI... Eventually the blocks we show that
contain info will become a reusable thing that we call programmatically.

But for now, we just shove it in an XML file and duplicate a bunch of
XML, for each piece of info we want to show.