Send the registration id to Sprunge for now.
When we switch to fedmsg-notifications and move everything to production
this will go away. Right now it's pretty hacky. You end up having to
set up the fedmsg-gcm-demo daemon locally, getting an "Sender ID" from the
Google API console, typing that into Fedora Mobile, and then getting a
"Registration ID" back in the form of a sprunge.us link. Once you have that,
you have to copy it into a YAML file.

This is hugely a proof of concept, just because I want to be able to demo this
stuff at Flock in August.