NewsfeedAdapter -> FedmsgAdapter, and move it.
In an attempt to clean up our activities and have reusable pieces where we
can, I'm going to move our list adapters out to an adapter directory.

This does that with the NewsfeedAdapter which has been renamed to
FedmsgAdapter, because it's less general than every possible newsfeed
we might one day have.