Fix crash when HRF/datagrepper is unavailable
This prevents us from fataling but potentially puts the PullToRefreshAttacher
into a state of never completing. We should put a timeout in place and mark
the refresh as "complete" after a few seconds so we don't render the main
newsfeed unresponsive after the first refrsh attempt.

Note that this doesn't lock up the entire app, but it makes the newsfeed
unusable until the app is restarted, if this condition happens.