A very first go at a Fedmsg counter widget.
This has a lot of issues, but it is a start. The most important issue is that
we probably want to move the call to Datagrepper out to a Service, so that we
can let the user decide how often it gets updated, instead of always updating
it every half-hour.